Setup Database (#27)
* Connecting to DB

* Updating .env Example

* Create Models

* Updating Models

* Adding Channel Partial

This allows the bot to respond to DMs

* Update Models

* Install Mongoose

* Test ValidateEnv

* Grabbing files from main

* Install Mongoose

+++ src/index.ts
@@ -1,5 +1,7 @@
 import {LogLevel, SapphireClient} from '@sapphire/framework';
 import {Intents} from 'discord.js';
+import {validateEnv} from './utils/validateEnv';
+import {connectDatabase} from './database/connectDatabase';
 
 require('dotenv').config({path: '../.env'});
 
@@ -10,13 +12,17 @@ const client = new SapphireClient({
   logger: {
     level: LogLevel.Info,
   },
+  partials: ['CHANNEL'],
 });
 
 const main = async () => {
+  if (!validateEnv()) return;
+
   try {
     client.logger.info('Bot is logging in');
     await client.login(process.env.BOT_TOKEN);
     client.logger.info('Bot has logged in');
+    await connectDatabase();
   } catch (error) {
     client.logger.fatal(error);
     client.destroy();
